add: warning of data loss if rewrite of updated checksum files failed
The KeyboardInterrupt exception cannot be re-raised in a destructor (where the
checksum file cleanup took place until now). Also, even though it worked in my
own runs, the desctructor may not always be called at the proper time for the
cleanup to take place.

Thus, the class was refactored to be used in a with-statement, so it will be
cleaned up when the with block ends.